Delete stack if creation (not update) fails.
CF leaves a stack in ROLLBACK_COMPLETED state in case a stack creation fails and the failure method is set to ROLLBACK. This is a final state and any subsequent stack create or update attempt will fail afterwards. The policy has to be set to DELETE to ensure a state that can be continued with.