Support Lambda@Edge functions
Lambda@Edge functions do not allow Environment Variables. This change
skips creation of the SERVERLESS_ALIAS env var for any functions listed
as Lambda@Edge functions in cfg below.

I'm certainly open to other implementation approaches, but it'd be great
to be able to use this plugin in combination with Lambda@Edge.

```
custom:
  lambdaEdgeFunctions:
    - platform-spa-origin-request-lambda-edge
```