Clarify install for less-experienced npm users
The serverless framework makes use of npm, but is designed to allow deployment to services such as AWS which support a variety of languages. As a result, there are users utilizing the serverless tool who are not as familiar with npm commands. 

This change adds the explicit command syntax needed to accomplish the installation to the README install instructions